centerècg should be an int

+++ Common/Context.h
@@ -91,7 +91,7 @@ public :
   float quaternion[4];        // the actual quaternion used for "trackball" rotating 
   int useTrackball;           // do or do not use the trackball for rotations 
   double rotation_center[3];  // point around which to rotate the scene
-  double rotation_center_cg;  // rotate around the center of gravity instead of rotation_center[]
+  int rotation_center_cg;     // rotate around the center of gravity instead of rotation_center[]
   double min[3];              // x, y and z min for the current geometry 
   double max[3];              // x, y and z max for the current geometry 
   double cg[3];               // "center of gravity" of the current geometry
